libs/authentication/src/service/password-validation.service.ts
import { Inject, Injectable } from '@nestjs/common';

import { AUTHENTICATION_MODULE_OPTIONS_TOKEN } from '../authentication.module-definition';
import { AuthenticationModuleOptions } from '../authentication.module-options';

@Injectable()
export class PasswordValidationService {
  constructor(
    @Inject(AUTHENTICATION_MODULE_OPTIONS_TOKEN)
    private readonly authenticationModuleOptions: AuthenticationModuleOptions,
  ) {}

  validatePassword(user: unknown, password: string): Promise<boolean> {
    return this.authenticationModuleOptions.callback.validatePassword(
      user,
      password,
    );
  }
}
